+++ src/main/wrap/header.js
@@ -1,0 +1,8 @@
+/**
+ * jQuery Mobile angularJS adaper v${project.version}
+ * http://github.com/tigbro/jquery-mobile-angular-adapter
+ *
+ * Copyright 2011, Tobias Bosch (OPITZ CONSULTING GmbH)
+ * Licensed under the MIT license.
+ */
+(function() {